pyproject.toml
[tool.poetry]
name = "dawn_of_the_worlds"
version = "0.1.0"
description = ""
authors = ["dmitrii <dima240669@mail.ru>"]

[tool.poetry.dependencies]
python = "^3.9"
numpy = "^1.22.0"
Pillow = "^9.0.0"
aiogram = "^2.18"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
